Added guard and simplecov to github-webhook
