+++ github-webhook/lib/github-webhook.rb
@@ -1,0 +1,3 @@
+module GithubWebhook
+	autoload :App, 'github-webhook/app'
+end
