+++ github-webhook/spec/spec_helper.rb
@@ -15,7 +15,13 @@
 #
 # See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 
-require_relative '../app'
+require 'cadre/rspec3'
+require 'simplecov'
+SimpleCov.start
+
+$LOAD_PATH.unshift File.expand_path(File.dirname(__FILE__), '../lib')
+
+require 'github-webhook'
 
 RSpec.configure do |config|
   # rspec-expectations config goes here. You can use an alternate
